a lot of changes, new web services added for user, customer, credit card, person , sidebar and tab, all layers some basic implementation have been completed.
Test cases remain and some config for email remains.
Todo: figure how to commit a template file for the email properties. email authentication that is very sensitive so cannot be committed.
Priority: Commit changes so the project doesn't depend on env variables for local testing and keep a prod file locally for prod